requirements.txt
# general modules
pyTelegramBotAPI==4.24.0
python-decouple
aiohttp

# hot realoding modules
# watchdog